pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "virtuoso-bridge"
version = "0.4.0"
description = "Lightweight Python bridge for Cadence Virtuoso SKILL execution and Spectre simulation"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
virtuoso-bridge = "virtuoso_bridge.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"virtuoso_bridge.resources" = [".env_template", "*.yaml"]
"virtuoso_bridge.virtuoso.basic.resources" = ["*.il", "*.py"]
